Move PreferencesDialog up to components
