+++ packages/studio-base/src/components/WorkspaceDialogs.tsx
@@ -2,7 +2,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { PreferencesDialog } from "@foxglove/studio-base/components/AppBar/Preferences/PreferencesDialog";
+import { PreferencesDialog } from "@foxglove/studio-base/components/PreferencesDialog";
 import {
   useWorkspaceActions,
   useWorkspaceStore,
